gacha-gen: move utility scripts into gacha_common
just to avoid cluttering root dir with stuff that shouldn't be used directly
