+++ gacha_common/gacha_data/limited_rotation.csv
