+++ gacha_common/gacha_data/permanent.csv
