+++ gacha_common/gen_gacha_banner_image.py
